+++ src/util/index.js
@@ -70,9 +70,8 @@ export function isInteger(i) {
          hval ^= str.charCodeAt(ii);
          hval += (hval << 1) + (hval << 4) + (hval << 7) + (hval << 8) + (hval << 24);
      }
-     
      hval = hval >>> 0;
-
+     
      if( asString ){
          // Convert to 8 digit hex string
          return ("0000000" + hval.toString(16)).substr(-8);