+++ src/util/map_entity_refs.js
@@ -23,8 +23,9 @@ export default function mapComponentEntityRefs( registry, component, entityIdMap
         return component;
     }
 
-    const componentSchema = registry.schemaRegistry.getSchema( component.getSchemaUri() );
-    const componentProperties = componentSchema.get('properties');
+    console.log('getting com def', component.getDefUri() );
+    const componentDef = registry.schemaRegistry.getComponentDef( component.getDefUri() );
+    const componentProperties = componentDef.get('properties');
     // properties = getProperties( registry.schemaRegistry, component.getSchemaUri() );